requirements_dev.txt
-r requirements.txt
ptvsd
pylint
autopep8
coverage
pre-commit
tox
pytest
flake8-django
pylint-django
black
pep8
coveralls